Add pragmas to avoid PEP8 warnings
These lines are too long according to the PEP8. But, I think we can
waive this occurrence as a special case. I think it's more useful to
keep the URLs as they are instead of splitting them by a backlash
because if they are kept as they are they can be easily identified by
grep when a need to update them arise.