Account for missing reference href attribute
The `href` attribute of the `reference` element is optional according to
the XCCDF specification. This commit solves the situation when the
`href` attribute is not set. In this situation, the reference source is
unknown, we don't know where this reference points to so we will put in
in a new "unknown" reference category to be displayed in the HTML
report.