Show collected items data
Currently, the "Show test details" section displays definition
of OVAL Object and OVAL State. This patch will add also displaying
OVAL Items collected from the system for the given OVAL Object.
This is a very popular feature of the old report which I think
will be useful also in the new report because it gives detailed
information about the actual state of the target.