Hide some remediation types
This change will cause that Anaconda and Kickstart remediation
types will not be shown in the HTML report. Anaconda remediations
are consumed only by OSCAP Anaconda Addon. Kickstart remediations
are consumed by OpenSCAP to generate a kickstart. It isn't useful
to the user at all to see or copy these remediation types, they
don't have any use for them. To reduce the unnecessary information
in the report, we won't include them in the report.