Remove old style HTML report
This feature only calls XSLT templates from the OpenSCAP project,
which are copied into this repository. The copy of these templates
is duplicate code, and isn't synchronized with OpenSCAP, so it
becomes outdated. The new report has been usable for multiple
upstream releases, therefore, this "old-style report" feature
doesn't make any sense and should be removed.