Fix colors in CPE AL details
If negation was used in OVAL check for a CPE platform, the
results were red even if the check passed. It was a problem
when processing the negation.

In this patch, we will change the logic so that the red and green
colors are inverted when negation is used.

Also fixes a typo: frac -> fact.

Fixes: #250